// back/src/modules/xContentCrawler/types.ts
  export interface ContentInsight {
    hasValue: boolean;
    category: 'trading_idea' | 'project_intro' | 'market_insight' | 'none';
    // keyDetails: string[];
    summary: string;
    source: string;
    username: string;
    agentId: string; // 添加 agentId 字段
    timestamp: string;
    entity: Entity;
    event: Event;
  }

  


  export interface XPostPayload {
  userId: string;
  agentId: string; // 添加 agentId 字段
  post_content: string;
  authorUsername: string;
  url?: string;
  timestamp?: string;
}


export interface Entity {
  name: string;
  context: string;
}
export interface Event {
  name: string;
  details: string;
}